+++ src/component/CreateInvoice/index.jsx
@@ -2,10 +2,21 @@ import { useEffect, useState } from "react"
 import Dropdown from "../ClientDropdown"
 import moment from "moment";
 import StatusIcon from "../StatusIcon";
+import InvoiceRow from "../InvoiceRow";
+import { Link } from "react-router-dom";
 
 function CreateInvoice() {
     const [formattedDateCreated, setFormattedDateCreated] = useState("")
     const [formattedDateDue, setFormattedDateDue] = useState("")
+
+    const [clients, setClients] = useState([])
+
+    const [details, setDetails] = useState([{
+        "quantity": 0,
+        "rate": 0,
+        "total": 0,
+        "description": "Optional text field"
+    }])
 
     useEffect(() => {
         const dateDue = moment().add(30, 'days')
@@ -14,38 +25,29 @@ function CreateInvoice() {
         setFormattedDateCreated(dateCreated.format("DD MMMM YYYY"))
     }, [])
 
-    const [clients, setClients] = useState([])
-    const [clientName, setClientName] = useState('')
-    const [clientStreet, setClientStreet] = useState('')
-    const [clientCity, setClientCity] = useState('')
-
     useEffect(() => {
         fetch('https://invoicing-api.dev.io-academy.uk/clients')
             .then(res => res.json())
             .then((clientsData) => {
                 setClients(clientsData.data)
-                setClientName(clients.name)
-                setClientStreet(clients.street_address)
-                setClientCity(clients.city)
-
             })
     }, [])
 
-    const [quantity, setQuantity] = useState(0)
-    const [rate, setRate] = useState(Number(0))
-
-    function handleQuantity(e) {
-        setQuantity(Number(e.target.value))
-
+    function addDetails() {
+        setDetails(details.push({
+            "quantity": 0,
+            "rate": 0,
+            "total": 0,
+            "description": "Optional text field"
+        }))
+        console.log(details)
     }
-    function handleRate(e) {
-        setRate(Number(e.target.value))
-
-    }
-
 
     return (
         <div className="bg-white">
+
+            <Link to="/CreateInvoice">+ New Invoice</Link>
+
             <div className="grid grid-cols-2 place-content-around p-3">
                 <div className="pl-3">
                     <p className="font-medium pb-2 pt-2">From</p>
@@ -71,18 +73,8 @@ function CreateInvoice() {
                 <p>Rate</p>
                 <p>Total</p>
             </div>
-            <div className="grid grid-cols-4">
-                <textarea className="border border-gray-300 m-2 p-2" placeholder='Description'></textarea>
-                <input className="border border-gray-300 m-2 p-2 h-1/3	" onChange={handleQuantity} type='number' min='1' max='100' placeholder='Quantity' step='1'></input>
-                <input className="border border-gray-300 m-2 p-2 h-1/3	" onChange={handleRate} type='number' min='1' max='100' placeholder='Rate' step='1'></input>
-
-                <div className="grid grid-cols-2">
-                    <p>£{parseFloat(parseFloat(quantity * rate).toFixed(2)).toLocaleString()}</p>
-                    <button className="p-2 m-1 text-white bg-green-400 rounded">+</button>
-                    <span></span>
-                    <button className="p-2 m-1 text-white bg-red-500 rounded">-</button>
-                </div>
-            </div>
+            {details.map(detail => <InvoiceRow addDetails={addDetails} />}
+            {console.log(details)}
 
         </div>
     )
