+++ src/component/Header/index.jsx
@@ -12,7 +12,7 @@ function Header() {
     }, [])
 
     return (
-        <header className="pt-8 ">
+        <header className="pt-8 inline-block">
             <h1 className="font-bold text-4xl pb-1">Invoices</h1>
             <p className="text-gray-500 pb-8 text-lg font-medium">There are {unpaid} unpaid invoices</p>
         </header>
